requirements.txt
flask==3.0.0
clickhouse-connect==0.7.0
python-dotenv==1.0.0
tqdm==4.66.1 